Create forum.html
This week you will need to build a simple form.  This exercise will combine our HTML skills and JavaScript skills.

The form will consists of the  following:

a text box where an individual's UCID can be type in
a submit button which the user will click to activate a JS script that will validate the input.
The script will test the input against a pattern (regular expressions must be used) and if it meets the criteria you will then need to verify the UCID against a list of valid UCIDs that you create.

If the UCID is valid but not verifiable you will output "VALID UCID FORMAT BUT INVALID UCID".  
If the UCID is valid and verifiable you will output "VALID UCID FORMAT AND UCIID FOUND" 
If not it will output "UCID DOES NOT CONFORM TO VALID FORMAT".
The UCID pattern is as follows:

a sequence of letters followed by zero to three digits

The input element in the form should have the following attributes:

type
id
name
placeholder  
title
url: https://web.njit.edu/~jcq2/forum.htm