Create password.html
This week you will need to create a HTML form that allows a user to enter a password.  You will also need to create a script that will either hide or show the user's password and will change the background color of the text box based upon whether the password is being displayed or hidden. Remember this assignment must be implemented using JS.

The script will use the getElementById() method to toggle the between the password being displayed or hidden and the background color change of the text box. The user will need to click the text box in order for the changes to occur. The initial setting for the password text box should be hidden and the password when typed in should remain hidden until the password is entered and the user clicks the box

url: https://web.njit.edu/~jcq2/password.html