Create email.html
assignment: This week you will need to create a JS script that uses jQuery that will examine and confirm the initial email entered by the user with the second email entered by the user.  Once the confirm button is clicked  your script should verify that the two emails are equal.  If they are not the user should be notified that the emails do not match.  If the do match an alert should be issued indicating that the email has been confirmed.  This exercise will combine our HTML skills and jQuery skills.

url: https://web.njit.edu/~jcq2/email.html